+++ src/module-2/lesson-5/FunctionEvents.sol
@@ -3,10 +3,10 @@ pragma solidity 0.8.30;
 
 contract Events {
     address private sender;
-    uint private balance;
+    uint256 private balance;
 
-    event LogInitialize(address _sender, uint _balance);
-    event LogIncreaseBalance(address _sender, uint _amount);
+    event LogInitialize(address _sender, uint256 _balance);
+    event LogIncreaseBalance(address _sender, uint256 _amount);
 
     constructor() payable {
         sender = msg.sender;
@@ -18,14 +18,14 @@ contract Events {
     function getSender() public view returns (address) {
         return sender;
     }
-    
+
     function getBalance() external view returns (uint256) {
         return balance;
     }
 
     function increaseBalance() external payable returns (bool) {
         balance += msg.value;
-        
+
         emit LogIncreaseBalance(msg.sender, msg.value);
 
         return true;
